--- client/src/app/parts/profile/profileActions.js
@@ -1,10 +1,0 @@
-import { createActionAsync} from 'redux-act-async';
-import ProfileResources from './profileResources';
-
-export default function(rest){
-    let profile = ProfileResources(rest);
-    return {
-        get: createActionAsync('PROFILE_GET', profile.get),
-        update: createActionAsync('PROFILE_UPDATE', profile.update)
-    }
-}
