--- client/src/app/parts/profile/profileResources.js
@@ -1,10 +1,0 @@
-export default function(rest){
-  return {
-    get() {
-        return rest.get('me');
-    },
-    update(payload) {
-        return rest.patch('me', payload);
-    }
-  }
-};
