--- client/src/app/parts/profile/reducers/reducers.js
@@ -1,8 +1,0 @@
-import {createReducerAsync} from 'redux-act-async';
-
-export default function(actions){
-  return {
-    profile: createReducerAsync(actions.get),
-    profileUpdate: createReducerAsync(actions.update),
-  }
-}
